fat protein carbohidrates made float

+++ type_one/ingredients/forms.py
@@ -37,9 +37,9 @@ class IngredientUnitForm(forms.Form):
 class IngredientForm(forms.ModelForm):
     name = forms.CharField(widget=forms.TextInput(attrs={'class' : 'form-control input-sm'}))
     glycemic_index = forms.IntegerField(widget=forms.NumberInput(attrs={'class' : 'form-control input-sm'}), required=False)
-    fat_per_100g = forms.IntegerField(widget=forms.NumberInput(attrs={'class' : 'form-control input-sm'}))
-    carbohydrate_per_100g = forms.IntegerField(widget=forms.NumberInput(attrs={'class' : 'form-control input-sm'}))
-    protein_per_100g = forms.IntegerField(widget=forms.NumberInput(attrs={'class' : 'form-control input-sm'}))
+    fat_per_100g = forms.FloatField(widget=forms.NumberInput(attrs={'class' : 'form-control input-sm'}))
+    carbohydrate_per_100g = forms.FloatField(widget=forms.NumberInput(attrs={'class' : 'form-control input-sm'}))
+    protein_per_100g = forms.FloatField(widget=forms.NumberInput(attrs={'class' : 'form-control input-sm'}))
     energy_kKkal_per_100g = forms.IntegerField(widget=forms.NumberInput(attrs={'class' : 'form-control input-sm'}))
     class Meta:
         model = models.Ingredient
